+++ packages/app/src/features/topbar/components/topbar-navigation/TopbarNavigationDialog.tsx
@@ -4,7 +4,7 @@ import { assets } from '@/ui/assets'
 import { Accordion, AccordionContent, AccordionItem, AccordionTrigger } from '@/ui/atoms/accordion/Accordion'
 import { Dialog, DialogContent, DialogTitle, DialogTrigger } from '@/ui/atoms/dialog/Dialog'
 import { IconButton } from '@/ui/atoms/icon-button/IconButton'
-import { MenuItem } from '@/ui/atoms/new/menu-item/MenuItem'
+import { MenuItem } from '@/ui/atoms/menu-item/MenuItem'
 import { cn } from '@/ui/utils/style'
 import { ChevronDownIcon } from 'lucide-react'
 import { useState } from 'react'
