+++ packages/app/src/ui/atoms/link/Link.stories.tsx
@@ -1,11 +1,11 @@
 import { Meta, StoryObj } from '@storybook/react'
 import { withRouter } from 'storybook-addon-remix-react-router'
-import { MenuItem } from '../new/menu-item/MenuItem'
+import { MenuItem } from '../menu-item/MenuItem'
 import { Link } from './Link'
 
 const meta: Meta<typeof Link> = {
   decorators: [withRouter()],
-  title: 'Components/Atoms/new/Link',
+  title: 'Components/Atoms/Link',
   component: Link,
 }
